+++ buildSrc/src/main/kotlin/dev/muon/dynamic_difficulty/gradle/Properties.kt
@@ -6,11 +6,11 @@ object Properties {
     const val ARCHIVES_NAME = "dynamic_difficulty"
     const val MOD_ID = "dynamic_difficulty"
     const val MOD_AUTHOR = "MuonR"
-    val MOD_CONTRIBUTORS = listOf("MuonR")
+    val MOD_CONTRIBUTORS = listOf("Daripher", "SHXRKIE")
     const val DESCRIPTION = "It's Leveling Time"
     const val LICENSE = "ARR"
 
-    const val HOMEPAGE = "https://modrinth.com/mod/dynamic-difficulty"
+    const val HOMEPAGE = "https://www.curseforge.com/minecraft/mc-mods/dynamic-difficulty-mod"
     const val CURSEFORGE_PROJECT_ID = "1272655"
     const val CURSEFORGE_PAGE = "https://www.curseforge.com/minecraft/mc-mods/dynamic-difficulty-mod"
     const val MODRINTH_PROJECT_ID = "x"